chat-assistant/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.12
python-dotenv>=1.0.0
sse-starlette>=1.8.2
pydantic>=2.5.0